requirements.txt
flask
pymupdf
sentence-transformers
faiss-cpu
spacy
langdetect
pandas
scikit-learn
numpy
